added liquibase, finished services, added dto and mapper
