+++ src/main/java/com/example/HotelApp/controller/FeedbackController.java
@@ -1,10 +1,7 @@
 package com.example.HotelApp.controller;
 
-import com.example.HotelApp.entity.Feedback;
-import com.example.HotelApp.repository.FeedbackRepository;
 import com.example.HotelApp.service.FeedbackService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
